requirements.txt
python-twitter
google-api-python-client
pysocks
requests
